chore(refactor): move exception error handler
move handle_exception error handler from base_validation module to main module

[Start #158183497]
